+++ pkg/message/l3-vpn.go
@@ -43,7 +43,7 @@ func (p *producer) l3vpn(nlri bgp.MPNLRI, op int, ph *bmp.PerPeerHeader, update 
 
 		if ases := update.BaseAttributes.ASPath; len(ases) != 0 {
 			// Last element in AS_PATH would be the AS of the origin
-			prfx.OriginAS = int32(ases[len(ases)-1])
+			prfx.OriginAS = ases[len(ases)-1]
 		}
 		if nlri.IsIPv6NLRI() {
 			// IPv6 specific conversions
